スプライト途中

+++ DirectXGame/Engine/Base/Utility/DxCreateLib.h
@@ -145,14 +145,14 @@ namespace DxCreateLib
 		inline static D3D12_CPU_DESCRIPTOR_HANDLE GetCPUDescriptorHandle(ID3D12DescriptorHeap* descriptorHeap, uint32_t descriptorSize, uint32_t index) 
 		{
 			D3D12_CPU_DESCRIPTOR_HANDLE handleCPU = descriptorHeap->GetCPUDescriptorHandleForHeapStart();
-			handleCPU.ptr += (descriptorSize * index);
+			handleCPU.ptr += ((static_cast<long long int>(descriptorSize)) * index);
 			return handleCPU;
 		}
 
 		inline static D3D12_GPU_DESCRIPTOR_HANDLE GetGPUDescriptorHandle(ID3D12DescriptorHeap* descriptorHeap, uint32_t descriptorSize, uint32_t index)
 		{
 			D3D12_GPU_DESCRIPTOR_HANDLE handleCPU = descriptorHeap->GetGPUDescriptorHandleForHeapStart();
-			handleCPU.ptr += (descriptorSize * index);
+			handleCPU.ptr += ((static_cast<long long int>(descriptorSize)) * index);
 			return handleCPU;
 		}
